refactor(name-field): use same name field for both jobs and sequences
