+++ src/components/FormFields/NameField.js
@@ -14,7 +14,7 @@ const { Field } = ReactFinalForm
 const FIELD_NAME = 'name'
 const VALIDATOR = composeValidators(string, hasValue)
 
-const SequenceNameField = () => (
+const NameField = () => (
     <Field
         name={FIELD_NAME}
         component={InputFieldFF}
@@ -25,4 +25,4 @@ const SequenceNameField = () => (
     />
 )
 
-export default SequenceNameField
+export default NameField
