+++ src/components/FormFields/index.js
@@ -5,11 +5,10 @@ const fieldNames = {
 }
 
 export { default as CronField } from './CronField'
-export { default as JobNameField } from './JobNameField'
 export { default as JobTypeField } from './JobTypeField'
+export { default as NameField } from './NameField'
 export { default as ParameterFields } from './ParameterFields'
 export { default as ScheduleField } from './ScheduleField'
-export { default as SequenceNameField } from './SequenceNameField'
 export { default as SequenceOrderField } from './SequenceOrderField'
 
 export { fieldNames }
