+++ src/components/Forms/JobAddForm.js
@@ -6,7 +6,7 @@ import { DiscardFormButton } from '../Buttons'
 import { FormErrorBox } from '../FormErrorBox'
 import {
     ScheduleField,
-    JobNameField,
+    NameField,
     JobTypeField,
     ParameterFields,
     fieldNames,
@@ -48,7 +48,7 @@ const JobAddForm = ({
     return (
         <form onSubmit={handleSubmit}>
             <Box maxWidth="600px">
-                <JobNameField />
+                <NameField />
             </Box>
             <Box marginTop="16px" maxWidth="400px">
                 <JobTypeField />
